Final Code for Prototype Tracker Turtle
This code operates the Tracker Turtle anti-theft device. This code allows the users smart phone to communicate with the Lightblue Bean embedded in the shell of the turtle. When the accelerometer on the Lightblue Bean detects movement, a buzzer will sound and LED's on the turtle will glow red. Additionally, a notification will be sent to the user's phone saying movement was detected, and their things are being stolen.